Første commit siden afleveringen. - Ændrede alle 64-bit typer til 32-bit typer, siden opløsningen af lydfiler typisk højest er 24-bits. - Tilføjede asynkrone metoder for at gøre programmet mere imødekommende. - For yderligere at gøre programmet "hurtigere" er der nu igen en limit på maks antal datapunkter der tegnes på tidsgrafen.

+++ Aud.IO/AudioFile.cs
@@ -42,12 +42,12 @@
         /// Returns the audio data in analog (demodulated from LPCM).
         /// </summary>
         /// <returns>LPCM demodulated audio.</returns>
-        public abstract double[] GetDemodulatedAudio();
+        public abstract float[] GetDemodulatedAudio();
 
         /// <summary>
         /// Sets audio by audio data not modulated using LPCM.
         /// </summary>
         /// <param name="audio">LPCM modulated audio.</param>
-        public abstract void SetDemodulatedAudio(double[] audio);
+        public abstract void SetDemodulatedAudio(float[] audio);
     }
 }
